requirements.txt
fastapi
pydantic
google-generativeai
rdflib
uvicorn